pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "audio-notes-automation"
version = "0.1.0"
description = "Transcribe audio, label speakers, and generate organized notes."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "faster-whisper>=1.2.0",
  "openai>=2.28.0",
  "PyYAML>=6.0.0",
  "typer>=0.12.0",
]

[project.optional-dependencies]
diarization = [
  "pyannote.audio>=3.3.0",
  "torchaudio>=2.4.0",
]

[project.scripts]
audio-notes-automation = "audio_notes_automation.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

